pyproject.toml
[tool.poetry]
name = "google-auth-test-api"
version = "0.1.0"
description = ""
authors = ["Sam Keen <sam.sjk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
authlib = "^1.3.2"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
starlette = "^0.38.2"
httpx = "^0.27.2"
itsdangerous = "^2.2.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
